+++ qcelemental/models/v1/procedures.py
@@ -75,6 +75,24 @@ class QCInputSpecification(ProtoModel):
     def _version_stamp(cls, v):
         return 1
 
+    def convert_v(
+        self, version: int
+    ) -> Union["qcelemental.models.v1.QCInputSpecification", "qcelemental.models.v2.AtomicSpecification"]:
+        """Convert to instance of particular QCSchema version."""
+        import qcelemental as qcel
+
+        if check_convertible_version(version, error="QCInputSpecification") == "self":
+            return self
+
+        dself = self.dict()
+        if version == 2:
+            dself.pop("schema_name")
+            dself.pop("schema_version")
+
+            self_vN = qcel.models.v2.AtomicSpecification(**dself)
+
+        return self_vN
+
 
 class OptimizationInput(ProtoModel):
     id: Optional[str] = None
@@ -297,6 +315,7 @@ class TorsionDriveInput(ProtoModel):
             return self
 
         dself = self.dict()
+        # dself = self.model_dump(exclude_unset=True, exclude_none=True)
         if version == 2:
             dself["input_specification"].pop("schema_version", None)
             dself["optimization_spec"].pop("schema_version", None)
@@ -363,6 +382,8 @@ class TorsionDriveResult(TorsionDriveInput):
                 k: [opthist_class(**res).convert_v(version) for res in lst]
                 for k, lst in dself["optimization_history"].items()
             }
+            # if dself["optimization_spec"].pop("extras", None):
+            #    pass
 
             self_vN = qcel.models.v2.TorsionDriveResult(**dself)